Number_guesser
The program generates a random number in the range from 1 to 100 and prompts the user to guess the number. If the user's guess is greater than the random number, the program should display the message 'Too much, try again.' If the guess is less than the random number, the program should display the message 'Too little, try again.' If the user correctly guesses the number, the program should congratulate them and display the message 'You guessed it, congratulations!